moved and renamed authentication base
